Fix parallel builds
When parallel builds occur there is the posibility that the same
assets will be triggerd on the same pooled page, this means that there
is a likely outcome that the resource will be intercepted and handled
twice, this is illegal with puppeteer.

Polling available pages up to 10 is also memory intensive for sequential
builds as 10 pages will be in memory. This can cause issues in low
memory enviroments.

- Remove pooling and open close pages on demand to reduce memory
footprint
- Open new page on demand to ensure that parallel builds that request
same resource do not conflict which causes the puppeteer to throw up
error that request is already handled.
- Add catch clauses around request interception to ensure that e2e tests
continue if there is an issue with percy.